% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{add_back_covariates}
\alias{add_back_covariates}
\title{Add back covariate effects to batch-corrected feature abundance data}
\usage{
add_back_covariates(adj_data, l_stand_feature, l_ind)
}
\arguments{
\item{adj_data}{feature-by-sample matrix of batch-adjusted feature abundances
(but without covariate effects), as returned by relocate_scale.}

\item{l_stand_feature}{list of per-feature standardization fits, as returned
by fit_stand_feature.}

\item{l_ind}{list of indicator matrices, as returned by construct_ind.}
}
\value{
feature-by-sample matrix of batch-adjusted feature abundances
with covariate effects retained.
}
\description{
Add back covariate effects to batch-corrected feature abundance data
}
