% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.batch.R
\name{adjust.batch}
\alias{adjust.batch}
\title{Zero-inflated empirical Bayes adjustment of batch effect in compositional feature
abundance data}
\usage{
adjust.batch(feature.abd, batch, covariates = NULL, data,
  zero.inflation = TRUE, pseudo.count = 0.5, diagnostics = FALSE,
  verbose = TRUE)
}
\arguments{
\item{feature.abd}{feature*sample matrix of feature abundance (counts preferred).}

\item{batch}{name of the batch variable.}

\item{covariates}{name(s) of additional covariates to adjust for
in the batch correction model.}

\item{data}{data frame of metadata, must contain batch and covariates (if present).}

\item{zero.inflation}{TRUE/FALSE for whether or not a zero-inflated model should be run
Default to TRUE (zero-inflated model). If set to FALSE then vanilla ComBat
(with parametric adjustment) will be performed.}

\item{pseudo.count}{pseudo count to add to the count table before log-transformation. Default
to 0.5. This should be changed for relative abundance (will automatically set to half of minimal
non-zero value if not specified).}

\item{diagnostics}{should diagnostic plots be generated? Deafault to FALSE.}

\item{verbose}{should verbose modelling information should be printed? Default to TRUE.}
}
\value{
feature x sample matrix of adjusted feature abundance.
}
\description{
Zero-inflated empirical Bayes adjustment of batch effect in compositional feature
abundance data
}
