% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{adjust_EB}
\alias{adjust_EB}
\title{Perform batch adjustment on standardized feature abundances, based on EB
shrinked per-batch location and scale parameters}
\usage{
adjust_EB(s_data, l_params_shrink, l_stand_feature, batchmod, n_batch,
  l_ind)
}
\arguments{
\item{s_data}{feature-by-sample matrix of standardized abundances.}

\item{l_params_shrink}{list of shrinked parameters, as returned by
fit_shrink.}

\item{l_stand_feature}{list of per-feature standardization fits, as returned
by fit_stand_feature.}

\item{batchmod}{design matrix for batch variables.}

\item{n_batch}{number of batches in the data.}

\item{l_ind}{list of indicator matrices, as returned by construct_ind.}
}
\value{
feature-by-sample matrix of batch-adjusted feature abundances.
}
\description{
Perform batch adjustment on standardized feature abundances, based on EB
shrinked per-batch location and scale parameters
}
\keyword{internal}
