% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{aprior}
\alias{aprior}
\title{EB prior estimation for scale parameters}
\usage{
aprior(delta_hat, na.rm = FALSE)
}
\arguments{
\item{delta_hat}{frequentist per-batch scale estimations.}

\item{na.rm}{whether or not missing values should be removed.}
}
\value{
shape hyper parameter
}
\description{
EB prior estimation for scale parameters
}
