% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AST}
\alias{AST}
\title{AST transformation (modified from Maaslin2 and is different)}
\usage{
AST(x)
}
\arguments{
\item{x}{vector of abundance to be transformed.}
}
\value{
transformed vector of abundance.
}
\description{
AST transformation (modified from Maaslin2 and is different)
}
\keyword{internal}
