% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{back_transform_abd}
\alias{back_transform_abd}
\title{Transform batch adjusted feature abundances back to the original scale in
feature_abd}
\usage{
back_transform_abd(adj_data, feature_abd, type_feature_abd)
}
\arguments{
\item{adj_data}{feature-by-sample matrix of batch-adjusted feature abundances
with covariate effects retained.}

\item{feature_abd}{original feature-by-sample matrix of abundances
(proportions or counts).}

\item{type_feature_abd}{type of feature abundance table (counts or
proportions). If counts, the final output will be rounded into counts as
well.}
}
\value{
feature-by-sample matrix of batch-adjusted feature abundances,
with covariate effects retained and scales consistent with original abundance
matrix.
}
\description{
Transform batch adjusted feature abundances back to the original scale in
feature_abd
}
\keyword{internal}
