% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{batch_adj}
\alias{batch_adj}
\title{Main function for batch effect adjustment}
\usage{
batch_adj(physeq, batch, adj_model = NULL)
}
\arguments{
\item{physeq}{phyloseq object with otu table and metadata containing variables
to adjust for in normalizing}

\item{batch}{batch variable}

\item{adj_model}{formula for covariates to adjust for}
}
\value{
a phyloseq object, with batch-corrected abundance data
}
\description{
Main function for batch effect adjustment
}
