% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{batch_adj_forSim}
\alias{batch_adj_forSim}
\title{Batch effect adjustment, only for simulation}
\usage{
batch_adj_forSim(physeq, batch, adj_model = NULL)
}
\arguments{
\item{physeq}{phyloseq object with otu table and metadata containing variables
to adjust for in normalizing}

\item{batch}{batch variable}

\item{adj_model}{formula for covariates to adjust for}
}
\value{
a list with normalized phyloseq object and simulation details
}
\description{
Batch effect adjustment, only for simulation
}
