% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{bprior}
\alias{bprior}
\title{EB prior estimation for scale parameters}
\usage{
bprior(delta_hat, na.rm = FALSE)
}
\arguments{
\item{delta_hat}{frequentist per-batch location estimations.}

\item{na.rm}{whether or not missing values should be removed.}
}
\value{
scale hyper parameter
}
\description{
EB prior estimation for scale parameters
}
