% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{catchToList}
\alias{catchToList}
\title{Utility for catching warning/error messages}
\usage{
catchToList(expr)
}
\arguments{
\item{expr}{an expression to run that can generate potential errors/warnings}
}
\value{
a list, capturing both the return value of the expression, as well
as generated erros/warnings (\code{NULL} if no errors/warnings)
}
\description{
Utility for catching warning/error messages
}
\keyword{internal}
