% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_batch}
\alias{check_batch}
\title{Check batch variable}
\usage{
check_batch(x, min_n_batch = 2)
}
\arguments{
\item{x}{batch variable.}

\item{min_n_batch}{min. number of batches (for MMUPHin functions to run).}
}
\value{
if no errors then the batch variables (factorized if not already)
}
\description{
Check batch variable
}
\keyword{internal}
