% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_maaslin_meta.R
\name{check_covariates_random}
\alias{check_covariates_random}
\title{Check random-effect covariates for clustering within batches}
\usage{
check_covariates_random(data_covariates, batch)
}
\arguments{
\item{data_covariates}{Data frame of random effect variables.}

\item{batch}{Factor vector of batch IDs.}
}
\value{
A logical matrix indicating if random effects have clustered 
        observations within each batch.
}
\description{
Check random-effect covariates for clustering within batches
}
\keyword{internal}
