% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_discrete_discover.R
\name{check_D}
\alias{check_D}
\title{Check dissimilarity object}
\usage{
check_D(D)
}
\arguments{
\item{D}{dissimilarity object.}
}
\value{
returns an error if D is not a dissimilarity. Otherwise D as a 
matrix.
}
\description{
Make sure that the input is a dissimilarity object
}
\keyword{internal}
