% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_lm_meta.R
\name{check_exposure}
\alias{check_exposure}
\title{Check exposure variable}
\usage{
check_exposure(exposure, batch)
}
\arguments{
\item{exposure}{exposure variable.}

\item{batch}{batch variable.}
}
\value{
vector of indicators per batch for whether or not the exposure can
be fitted within the batches
}
\description{
Check exposure variable
}
\keyword{internal}
