% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_metadata}
\alias{check_metadata}
\title{Check that metadata data frame has all the variables and not missing}
\usage{
check_metadata(data, variables, no_missing = TRUE)
}
\arguments{
\item{data}{data frame of metadata.}

\item{variables}{name of variables (batch, covariates, etc.) to check}
}
\value{
data reduced to include only those specified in variables
}
\description{
Check that metadata data frame has all the variables and not missing
}
