% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_options}
\alias{check_options}
\title{Utility for checking options}
\usage{
check_options(x, x_name, options)
}
\arguments{
\item{x}{the specified value}

\item{x_name}{name of the specified value}

\item{options}{allowed options}
}
\value{
error if \code{x} is not in \code{options}. Otherwise returns
\code{x}.
}
\description{
Utility for checking options
}
\keyword{internal}
