% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_options_continuous}
\alias{check_options_continuous}
\title{Utility for checking continuous options}
\usage{
check_options_continuous(x, x_name, range)
}
\arguments{
\item{x}{the specified numeric value}

\item{x_name}{name of the specified value}

\item{range}{allowed range}
}
\value{
error if \code{x} is not within \code{range} (boundaries 
excluded). Otherwise returns \code{x}.
}
\description{
Utility for checking continuous options
}
\keyword{internal}
