% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_pseudo_count}
\alias{check_pseudo_count}
\title{Utility for checking pseudo count}
\usage{
check_pseudo_count(x)
}
\arguments{
\item{x}{the specified pseudo count}
}
\value{
error if pseudo count is smaller than zero. Otherwise returns 
\code{x}.
}
\description{
Utility for checking pseudo count
}
\keyword{internal}
