% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{check_rank}
\alias{check_rank}
\title{Check if a design matrix is full rank}
\usage{
check_rank(design)
}
\arguments{
\item{design}{design matrix.}
}
\value{
TRUE/FALSE for whether or not the design matrix is full rank.
}
\description{
Check if a design matrix is full rank
}
\keyword{internal}
