% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_samples}
\alias{check_samples}
\title{Check that sample numbers and names match between a feature table and a
metadata data frame}
\usage{
check_samples(feature_abd, data)
}
\arguments{
\item{feature_abd}{feature-by-sample matrix of abundances (proportions or
counts).}

\item{data}{data frame of metadata.}
}
\value{
matched sample names
}
\description{
Sample names (column names of the feature table, row names of the metadata
data frame) must be matching exactly. Note that this dictates that they
cannot be NULL because by design data (a data frame) should have non-empty
row names.
}
\keyword{internal}
