% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_discrete_discover.R
\name{check_samples_D}
\alias{check_samples_D}
\title{Check that sample numbers and names match between a dissimilarity matrix and 
a metadata data frame}
\usage{
check_samples_D(D, data)
}
\arguments{
\item{D}{sample-by-sample matrix of dissimilarities (proportions or
counts).}

\item{data}{data frame of metadata.}
}
\value{
matched sample names
}
\description{
Sample names (row/column names of the D matrix, row names of the metadata
data frame) must be matching exactly. Note that this dictates that they
cannot be NULL because by design data (a data frame) should have non-empty
row names.
}
\keyword{internal}
