% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{clust_strength}
\alias{clust_strength}
\title{Wrapper function for unsupervised clustering strength evaluation}
\usage{
clust_strength(physeq, clust_method = "pam", dist_measure = "bray", kmax,
  plots)
}
\arguments{
\item{physeq}{phyloseq object with otu table and metadata containing variables
to adjust for in normalizing}

\item{clust_method}{clustering method to be used}

\item{dist_measure}{distance metric to be used}

\item{kmax}{max of number of clusters (groups) to be used}

\item{plots}{whether or not visualization plots should be generated}
}
\value{
a data frame summarizing clustering strength evaluations
}
\description{
Wrapper function for unsupervised clustering strength evaluation
}
