% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ComBat_forSim}
\alias{ComBat_forSim}
\title{Running original ComBat model, only for simulation}
\usage{
ComBat_forSim(dat, batch, mod = NULL, mean.only = FALSE,
  BPPARAM = bpparam("SerialParam"))
}
\description{
Running original ComBat model, only for simulation
}
