% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{ComBat_mod}
\alias{ComBat_mod}
\title{Running modified ComBat model}
\usage{
ComBat_mod(dat, batch, mod = NULL, mean.only = FALSE,
  BPPARAM = bpparam("SerialParam"))
}
\description{
Running modified ComBat model
}
