% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ComBat_mod_forSim}
\alias{ComBat_mod_forSim}
\title{Running modified ComBat model, only for simulation}
\usage{
ComBat_mod_forSim(dat, batch, mod = NULL, mean.only = FALSE,
  BPPARAM = bpparam("SerialParam"))
}
\description{
Running modified ComBat model, only for simulation
}
