% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{construct_ind}
\alias{construct_ind}
\title{Create indicator matrices for which feature/batch/samples to adjust. This is
relevant for zero_inflation is TRUE and only non-zero values are adjusted.}
\usage{
construct_ind(feature_abd, n_batch, design, zero_inflation)
}
\arguments{
\item{feature_abd}{feature-by-sample matrix of abundances (proportions or
counts).}

\item{n_batch}{number of batches in the data.}

\item{design}{design matrix.}

\item{zero_inflation}{zero inflation flag.}
}
\value{
list of indicator matrices needed by fitting in adjust_batch.
}
\description{
Create indicator matrices for which feature/batch/samples to adjust. This is
relevant for zero_inflation is TRUE and only non-zero values are adjusted.
}
