% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.discover.R
\name{continuous.discover}
\alias{continuous.discover}
\title{Main function for continuous structure discovery}
\usage{
continuous.discover(feature.count, batch, data, transform = "AST",
  var.perc.cutoff = 0.8, cor.cutoff = 0.707, directory = "./",
  diagnostics = TRUE, verbose = TRUE)
}
\arguments{
\item{feature.count}{Feature x sample matrix of feature abundance}

\item{batch}{Name of batch variable}

\item{data}{data frame}

\item{transform}{transformation}

\item{var.perc.cutoff}{percentage variance cutoff for top PCs}

\item{cor.cutoff}{correlation cutoff for constructing graph}

\item{directory}{output directory for figures}

\item{diagnostics}{logical. Should diagnostics be plotted?}

\item{verbose}{logical. Verbose?}
}
\value{
A list of MMUPHin continuous score discovery object
}
\description{
Main function for continuous structure discovery
}
