% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.train.R
\name{continuous.train}
\alias{continuous.train}
\title{Main function for batch effect adjustment}
\usage{
continuous.train(feature.count, batch, pseudo.count = 0.5,
  var.perc.cutoff = 0.3, cor.cutoff = NULL, directory = "./",
  diagnostics = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{feature.count}{Feature x sample matrix of feature abundance}

\item{batch}{Variable indicating batch membership}

\item{pseudo.count}{Pseudo count to add to the count table before log-transformation. Default
to 0.5.}
}
\value{
A list of MMUPHin continuous score discovery training object
}
\description{
Main function for batch effect adjustment
}
