% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_lm_meta.R
\name{create.table.Maaslin}
\alias{create.table.Maaslin}
\title{Utility for generating empty Maaslin2 results table}
\usage{
create.table.Maaslin(features, exposure, lvl.exposure)
}
\arguments{
\item{features}{name of the features fitted to Maaslin2.}

\item{exposure}{the exposure variable.}

\item{lvl.exposure}{levels of the exposure variable, if a factor.}
}
\value{
a table for each feature-exposure value pai; reference level of exposure, if a factor,
is taken out because is absorbed into the intercept term in Maaslin2 regression.
}
\description{
Utility for generating empty Maaslin2 results table
}
