% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_maaslin_meta.R
\name{create_table_maaslin}
\alias{create_table_maaslin}
\title{Generate dummy grid for Maaslin results table}
\usage{
create_table_maaslin(features, exposure, lvl_exposure)
}
\description{
Generate dummy grid for Maaslin results table
}
\keyword{internal}
