% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_continuous_discover.R
\name{diagnostic_continuous_discover}
\alias{diagnostic_continuous_discover}
\title{Diagnostic visualization for continuous.discover function}
\usage{
diagnostic_continuous_discover(mat_vali, lvl_batch, cos_cutoff, output)
}
\arguments{
\item{mat_vali}{matrix of maximum correlations between the cluster-specific 
consensus loadings and top PC loadings from each batch}

\item{lvl_batch}{unique batches in the data}

\item{cos_cutoff}{the specified consine coefficient cutoff}

\item{output}{output file name}
}
\value{
the invisble ggplot2 plot object
}
\description{
Diagnostic visualization for continuous.discover function
}
\keyword{internal}
