% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_discrete_discover.R
\name{diagnostic_discrete_discover}
\alias{diagnostic_discrete_discover}
\title{Diagnostic visualization for discrete.discover function}
\usage{
diagnostic_discrete_discover(stats_internal, stats_external, lvl_batch, output)
}
\arguments{
\item{stats_internal}{list of internal evaluation summary statistics}

\item{stats_external}{list of external validation summary statistics}

\item{lvl_batch}{unique batches in the data}

\item{output}{}
}
\value{
the invisble ggplot2 plot object
}
\description{
Diagnostic visualization for discrete.discover function
}
\keyword{internal}
