% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{diagnostics.adjust.batch}
\alias{diagnostics.adjust.batch}
\title{Diagnostic visualization for adj.batch function}
\usage{
diagnostics.adjust.batch(feature.abd, feature.abd.adj, batch, gamma.hat,
  gamma.star)
}
\arguments{
\item{feature.abd}{feature*sample matrix of unadjusted abundance}

\item{feature.abd.adj}{feature*sample matrix of adjusted abundance}

\item{batch}{vector of batch variables}

\item{gamma.hat}{estimated per feature-batch gamma parameters}

\item{gamma.star}{shrinked per feature-batch gamma parameters}
}
\value{
the invisble ggplot2 plot object
}
\description{
Diagnostic visualization for adj.batch function
}
