% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{diagnostics.continuous.discover}
\alias{diagnostics.continuous.discover}
\title{Title}
\usage{
diagnostics.continuous.discover(mat.vali, lvl.batch)
}
\arguments{
\item{mat.vali}{matrix of maximum correlations between the cluster-specific consensus loadings and
top PC loadings from each batch}

\item{lvl.batch}{unique batches in the data}
}
\value{
the invisble ggplot2 plot object
}
\description{
Title
}
