% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{diagnostics.discrete.discover}
\alias{diagnostics.discrete.discover}
\title{Diagnostic visualization for discrete.discover function}
\usage{
diagnostics.discrete.discover(stats.internal, stats.external, lvl.batch)
}
\arguments{
\item{stats.internal}{list of internal evaluation summary statistics}

\item{stats.external}{list of external validation summary statistics}

\item{lvl.batch}{unique batches in the data}
}
\value{
the invisble ggplot2 plot object
}
\description{
Diagnostic visualization for discrete.discover function
}
