% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{diagnostics_adjust_batch}
\alias{diagnostics_adjust_batch}
\title{Diagnostic visualization for adj_batch function}
\usage{
diagnostics_adjust_batch(feature_abd, feature_abd_adj, batch, gamma_hat,
  gamma_star, output)
}
\arguments{
\item{feature_abd}{original feature-by-sample matrix of abundances
(proportions or counts).}

\item{feature_abd_adj}{ffeature-by-sample matrix of batch-adjusted feature
abundances, with covariate effects retained and scales consistent with
original abundance matrix.}

\item{batch}{the batch variable (should be a factor).}

\item{gamma_hat}{estimated per feature-batch gamma parameters.}

\item{gamma_star}{shrinked per feature-batch gamma parameters}
}
\value{
(invisbly) the ggplot2 plot object
}
\description{
Diagnostic visualization for adj_batch function
}
