% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.discover.R
\name{discrete.discover}
\alias{discrete.discover}
\title{Main function for discrete structure discovery}
\usage{
discrete.discover(feature.count, batch, data, distance = "bray",
  k.max = 10, cluster.method = fpc::claraCBI,
  classify.method = "centroid", M = 10, nnk = 1,
  diagnostics = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{feature.count}{Feature x sample matrix of feature abundance}

\item{batch}{Name of batch variable}

\item{data}{data frame}

\item{distance}{dissimilarity measure or dist object}

\item{k.max}{max k for testing}

\item{cluster.method}{method for clustering. consistent with fpc}

\item{classify.method}{method for classification. consistent with fpc}

\item{M}{number of folds validations during internal validation}

\item{nnk}{number of k nearest neighbors, if method is KNN}

\item{diagnostics}{logical. Should diagnostics be plotted?}

\item{verbose}{logical. Verbose?}

\item{...}{additional parameters for clsutering. consistent with fpc}
}
\value{
A list of MMUPHin continuous score discovery training object
}
\description{
Main function for discrete structure discovery
}
