% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.structure.R
\name{discrete.structure}
\alias{discrete.structure}
\title{Wrapper function for unsupervised clustering strength evaluation}
\usage{
discrete.structure(feature.count, batch, clst.method = "pam",
  diss.measure = "bray", kmax, plots)
}
\arguments{
\item{kmax}{max of number of clusters (groups) to be used}

\item{plots}{whether or not visualization plots should be generated}

\item{physeq}{phyloseq object with otu table and metadata containing variables
to adjust for in normalizing}

\item{clust_method}{clustering method to be used}

\item{dist_measure}{distance metric to be used}
}
\value{
a data frame summarizing clustering strength evaluations
}
\description{
Wrapper function for unsupervised clustering strength evaluation
}
