% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseDOSSA_helpers.R
\name{extract_sparseDOSSA}
\alias{extract_sparseDOSSA}
\title{Title}
\usage{
extract_sparseDOSSA(sparseDOSSA_fit, add_libsize_var = FALSE)
}
\arguments{
\item{add_libsize_var}{}
}
\value{

}
\description{
Title
}
