% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fill_dimnames}
\alias{fill_dimnames}
\title{Fill in artificial row/column names to a matrix or data frame, if they are
missing}
\usage{
fill_dimnames(x, row_prefix, col_prefix)
}
\arguments{
\item{x}{matrix or data frame}

\item{row_prefix}{prefix for the artificial row names}

\item{col_prefix}{prefix for the artificial column names}
}
\value{
x but with the missing dimention names filled in
}
\description{
Fill in artificial row/column names to a matrix or data frame, if they are
missing
}
