% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{fit_EB}
\alias{fit_EB}
\title{Parametric estimation of per-batch location and scale parameters, and
Empirical Bayes estimation of their priors}
\usage{
fit_EB(s_data, l_stand_feature, batchmod, n_batch, l_ind)
}
\arguments{
\item{s_data}{feature-by-sample matrix of standardized abundances.}

\item{l_stand_feature}{list of per-feature standardization fits, as returned
by fit_stand_feature.}

\item{batchmod}{design matrix for batch variables.}

\item{n_batch}{number of batches in the data.}

\item{l_ind}{list of indicator matrices, as returned by construct_ind.}
}
\value{
list of parameter estimations.
}
\description{
Parametric estimation of per-batch location and scale parameters, and
Empirical Bayes estimation of their priors
}
\keyword{internal}
