% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{fit_shrink}
\alias{fit_shrink}
\title{A posteriori shrink per-batch location and scale parameters towards their EB
priors}
\usage{
fit_shrink(s_data, l_params, batchmod, n_batch, l_ind, control)
}
\arguments{
\item{s_data}{feature-by-sample matrix of standardized abundances.}

\item{l_params}{list of parameter fits, as returned by fit_EB.}

\item{batchmod}{design matrix for batch variables.}

\item{n_batch}{number of batches in the data.}

\item{l_ind}{list of indicator matrices, as returned by construct_ind.}

\item{control}{list of control parameters (passed on to it_sol)}
}
\value{
list of shrinked per-batch location and scale parameters.
}
\description{
A posteriori shrink per-batch location and scale parameters towards their EB
priors
}
