% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{it.sol_mod}
\alias{it.sol_mod}
\title{Modified iterated solving function from sva, used for modified ComBat}
\usage{
it.sol_mod(sdat, g.hat, d.hat, g.bar, t2, conv = 1e-04)
}
\description{
Modified iterated solving function from sva, used for modified ComBat
}
