% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.lm.R
\name{lm.meta}
\alias{lm.meta}
\title{Main function for batch effect adjustment}
\usage{
lm.meta(feature.count, batch, exposure, formula.adj = NULL, data.adj = NULL,
  method = "REML", zero.inflation = TRUE, pseudo.count = 0.5,
  forestplots = TRUE, directory = "./", verbose = TRUE, ...)
}
\arguments{
\item{feature.count}{Feature x sample matrix of feature abundance}

\item{batch}{Variable indicating batch membership}

\item{exposure}{Primary exposure of interest}

\item{formula.adj}{Formula indicating additional covariate to adjust for
in the batch correction model}

\item{data.adj}{Data frame for covarites to adjust for.}

\item{zero.inflation}{Flag for whether or not a zero-inflated model should be run.
Default to TRUE (zero-inflated model). If set to FALSE then vanilla ComBat
(with parametric adjustment) will be performed.}

\item{pseudo.count}{Pseudo count to add to the count table before log-transformation. Default
to 0.5.}

\item{directory}{Directory for generated forest plots}

\item{verbose}{Flag for whether or not verbose modelling information should be printed.
Default to yes.#'}

\item{forest}{Flag for whether or not forest plots figures should be generated.
Deafault to TRUE.}
}
\value{
a long-format data.frame of main exposure's individual and aggregated effect sizes
}
\description{
Main function for batch effect adjustment
}
