% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_meta.R
\name{lm_meta}
\alias{lm_meta}
\title{Covariate-adjusted meta-analysis of per-feature associations in compositional data}
\usage{
lm_meta(feature_abd, exposure, batch, covariates = NULL,
  covariates_random = NULL, data, control)
}
\arguments{
\item{feature_abd}{feature*sample matrix of feature abundance (counts
preferred).}

\item{exposure}{name of the exposure of interest variable.}

\item{batch}{name of the batch variable.}

\item{covariates}{name(s) of additional covariates to adjust for in the
Maaslin2 model.}

\item{covariates_random}{name(s) of random covariates in the Maaslin2 model.}

\item{data}{data frame of metadata, must contain exposure, batch, covariates
(if present), and covariates_random (if present).}

\item{control}{}
}
\value{
a list with component meta.results for per-feature meta-analysis results, and
component l.Maaslin.fit which itself is a list of results from fitting Maaslin2 in
individual studies.
}
\description{
Covariate-adjusted meta-analysis of per-feature associations in compositional data
}
