% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_lm_meta.R
\name{Maaslin2.wrapper}
\alias{Maaslin2.wrapper}
\title{Wrapper function for Maaslin2}
\usage{
Maaslin2.wrapper(feature.abd, data, exposure, covariates = NULL,
  covariates.random = NULL, output = "./", normalization = "TSS",
  transform = "AST", analysis_method = "LM")
}
\arguments{
\item{feature.abd}{feature*sample matrix of feature abundance.}

\item{data}{data frame of metadata.}

\item{exposure}{name of exposure variable.}

\item{covariates}{name of covariates.}

\item{covariates.random}{name of random covariates.}

\item{output}{directory for Maaslin2.}

\item{normalization}{normalization parameter for Maaslin2.}

\item{transform}{transformation parameter for Maaslin2.}

\item{analysis_method}{analysis method parameter for Maaslin2.}
}
\value{
a data frame recording per-feature coefficients, p-values, etc. from running Maaslin2.
}
\description{
Wrapper function for Maaslin2
}
