% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_lm_meta.R
\name{Maaslin2_wrapper}
\alias{Maaslin2_wrapper}
\title{Wrapper function for Maaslin2}
\usage{
Maaslin2_wrapper(feature_abd, data, exposure, covariates = NULL,
  covariates_random = NULL, output = tempdir(),
  normalization = "TSS", transform = "AST", analysis_method = "LM")
}
\arguments{
\item{feature_abd}{feature*sample matrix of feature abundance.}

\item{data}{data frame of metadata.}

\item{exposure}{name of exposure variable.}

\item{covariates}{name of covariates.}

\item{covariates_random}{name of random covariates.}

\item{output}{directory for Maaslin2.}

\item{normalization}{normalization parameter for Maaslin2.}

\item{transform}{transformation parameter for Maaslin2.}

\item{analysis_method}{analysis method parameter for Maaslin2.}
}
\value{
a data frame recording per-feature coefficients, p-values, etc. from 
running Maaslin2.
}
\description{
Wrapper function for Maaslin2
}
\keyword{internal}
