% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_maaslin_meta.R
\name{Maaslin3_wrapper}
\alias{Maaslin3_wrapper}
\title{Internal wrapper for Maaslin 3 to ensure name parity and handle errors}
\usage{
Maaslin3_wrapper(
  feature_abd,
  data,
  exposure,
  covariates = NULL,
  covariates_random = NULL,
  output = tempdir(),
  control
)
}
\arguments{
\item{feature_abd}{Feature-by-sample matrix.}

\item{data}{Metadata data frame.}

\item{exposure}{Character; primary exposure variable.}

\item{covariates}{Character vector; fixed effects.}

\item{covariates_random}{Character vector; random effects.}

\item{control}{List of maaslin3 parameters.}
}
\value{
The standard Maaslin 3 fit object with restored original names.
}
\description{
Internal wrapper for Maaslin 3 to ensure name parity and handle errors
}
\keyword{internal}
