% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{match_control}
\alias{match_control}
\title{Match user-specified control parameters with default, and modify if needed}
\usage{
match_control(default, control)
}
\arguments{
\item{default}{list of default control parameters}

\item{control}{list of user-provided control parameters}
}
\value{
list of control parameters, set to user provided values if specified
and default other wise
}
\description{
Match user-specified control parameters with default, and modify if needed
}
\keyword{internal}
