% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{meta_fit}
\alias{meta_fit}
\title{Main function for batch effect adjustment}
\usage{
meta_fit(physeq, batch, fit_model = NULL, method, dir_output = "./")
}
\arguments{
\item{physeq}{combined phyloseq object with otu table and metadata across studies}

\item{fit_model}{formula for model to fit with Maaslin}

\item{method}{fit method for rma}

\item{dir_output}{output directory for Maaslin fit results}

\item{study}{study indicator variables}
}
\value{
a phyloseq object, with batch-corrected abundance data
}
\description{
Main function for batch effect adjustment
}
