% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Wrapper for different ways of data normalization}
\usage{
normalize(physeq, method = "none", rarefy_size = 5000)
}
\arguments{
\item{physeq}{phyloseq object to import}

\item{method}{method for normalization}

\item{rarefy_size}{the library size used for rarefying. Only meaningful if method = 'Rarefy'}
}
\value{
phyloseq object with normalized abundance table
}
\description{
Wrapper for different ways of data normalization
}
