% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalizeFeatures}
\alias{normalizeFeatures}
\title{Normalize feature abundance table}
\usage{
normalizeFeatures(features, normalization = "NONE", pseudo.count = 0)
}
\arguments{
\item{features}{feature*sample abundance table.}

\item{normalization}{normalization method.}

\item{pseudo.count}{pseudo count to be added to abundances before normalization.}
}
\value{
normalized feature*sample abundance table.
}
\description{
Normalize feature abundance table
}
