% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PERMANOVA}
\alias{PERMANOVA}
\title{Wrapper for performing PERMANOVA on microbial compositional data}
\usage{
PERMANOVA(physeq, dist = "bray", model, nperm = 99)
}
\arguments{
\item{nperm}{number of permutations to use in PERMANOVA}

\item{physeq_before}{phyloseq object before adjustment}

\item{physeq_after}{phyloseq object after adjustment}

\item{batch_var}{batch variable used for correction}

\item{adj_var}{additional variables to adjust for}

\item{distance}{distance metric used for PERMANOVA}
}
\value{
fitted PERMANOVA result
}
\description{
Wrapper for performing PERMANOVA on microbial compositional data
}
