% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postmean_mod}
\alias{postmean_mod}
\title{Modified posterior mean function from sva, used for modified ComBat}
\usage{
postmean_mod(g.hat, g.bar, n, d.star, t2)
}
\description{
Modified posterior mean function from sva, used for modified ComBat
}
