% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{postvar_mod}
\alias{postvar_mod}
\title{Modified posterior var function from sva, used for modified ComBat}
\usage{
postvar_mod(sum2, n, a, b)
}
\description{
Modified posterior var function from sva, used for modified ComBat
}
