% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{relocate_scale}
\alias{relocate_scale}
\title{Relocate and scale feature abundances to correct for batch effects, given
shrinked per-batch location and scale parameters}
\usage{
relocate_scale(s_data, l_params_shrink, batchmod, n_batch, l_ind)
}
\arguments{
\item{s_data}{feature-by-sample matrix of standardized abundances.}

\item{l_params_shrink}{list of shrinked parameters, as returned by
fit_shrink.}

\item{batchmod}{design matrix for batch variables.}

\item{n_batch}{number of batches in the data.}

\item{l_ind}{list of indicator matrices, as returned by construct_ind.}
}
\value{
feature-by-sample matrix of batch-adjusted feature abundances
(but without covariate effects).
}
\description{
Relocate and scale feature abundances to correct for batch effects, given
shrinked per-batch location and scale parameters
}
\keyword{internal}
