% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_lm_meta.R
\name{rename.Maaslin}
\alias{rename.Maaslin}
\title{Utility for temporarily renaming samples/features for Maaslin2 run to bypass the rare
cases where unconventional names can cause exceptions}
\usage{
rename.Maaslin(old.names, prefix)
}
\arguments{
\item{old.names}{vector of names.}

\item{prefix}{prefix for the replacement (new numbered names).}
}
\value{
vector of new names - numbered vector with same length as old names and with the specified
prefix.
}
\description{
Utility for temporarily renaming samples/features for Maaslin2 run to bypass the rare
cases where unconventional names can cause exceptions
}
