% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_maaslin_meta.R
\name{rename_maaslin}
\alias{rename_maaslin}
\title{Temporary renaming of samples/features for formula safety}
\usage{
rename_maaslin(old_names, prefix)
}
\description{
Temporary renaming of samples/features for formula safety
}
\keyword{internal}
