% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rma.mod.wrapper}
\alias{rma.mod.wrapper}
\title{Wrapper for fitting rma models with a moderator parameter. This allows to
analyze an interaction model with meta-analysis effects}
\usage{
rma.mod.wrapper(l.Maaslin.fit, data.moderator, method = "REML")
}
\arguments{
\item{l.Maaslin.fit}{list of Maaslin2 result data frames, outputted from Maaslin2.wrapper.}

\item{data.moderator}{data frame recording the moderator variables. Each row corresponds to
a single study, and should have the same number of rows as l.Maaslin.fit.}

\item{method}{meta-analysis model to run, options provided in metafor::rma.}
}
\value{
a data frame recording per-feature/moderator value meta-analysis association results.
(coefficients, p-values, etc.)
}
\description{
Wrapper for fitting rma models with a moderator parameter. This allows to
analyze an interaction model with meta-analysis effects
}
