% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rma.wrapper}
\alias{rma.wrapper}
\title{Wrapper for fitting fixed/random effects meta-analysis model using metafor}
\usage{
rma.wrapper(l.Maaslin.fit, method = "REML", forest.plots = TRUE,
  output)
}
\arguments{
\item{l.Maaslin.fit}{list of Maaslin2 result data frames, outputted from Maaslin2.wrapper.}

\item{method}{meta-analysis model to run, options provided in metafor::rma.}

\item{forest.plots}{should forest plots be generated (for the significant associations).}

\item{output}{directory for the output forest plots.}
}
\value{
a data frame recording per-feature meta-analysis association results.
(coefficients, p-values, etc.)
}
\description{
Wrapper for fitting fixed/random effects meta-analysis model using metafor
}
