% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_maaslin_meta.R
\name{rma_wrapper}
\alias{rma_wrapper}
\title{Wrapper for fitting fixed/random effects meta-analysis model using metafor}
\usage{
rma_wrapper(maaslin_fits, stream = "abundance", control)
}
\arguments{
\item{maaslin_fits}{list of results from individual batches.}

\item{stream}{stream to aggregate ("abundance" or "prevalence").}

\item{control}{control parameters including output path and plot settings.}
}
\value{
a data frame recording per-feature meta-analysis association results.
}
\description{
Wrapper for fitting fixed/random effects meta-analysis model using metafor
}
\keyword{internal}
