% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_pseudo}
\alias{set_pseudo}
\title{Set pseudo count for an abundance matrix. Pseudo count is currently set to
half of minimum non-zero values}
\usage{
set_pseudo(features)
}
\arguments{
\item{features}{feature-by-sample matrix of abundances (proportions or
counts).}
}
\value{
the pseudo count
}
\description{
Set pseudo count for an abundance matrix. Pseudo count is currently set to
half of minimum non-zero values
}
\keyword{internal}
