% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_NB}
\alias{simulate_NB}
\title{Simulate microbial read counts data with
negative binomial distribution}
\usage{
simulate_NB(read_depth = 10000, number_feature, number_samples, UserMetadata,
  spikeStrength = 1, noZeroInflate = FALSE, prob_spiked, verbose = FALSE)
}
\arguments{
\item{read_depth}{Average read depth}

\item{number_feature}{Number of simulated features}

\item{number_samples}{Number of simulated samples}

\item{UserMetadata}{A data frame with user-specified metadata}

\item{spikeStrength}{Magnitude of association between metadata and microbial features}

\item{noZeroInflate}{Whether zero-inflation should be simulated}

\item{prob_spiked}{Percentage of feature associated with metadata}

\item{verbose}{}
}
\value{
A phyloseq object with simulated read counts and
user-specified metadata
}
\description{
Simulate microbial read counts data with
negative binomial distribution
}
