% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transform_features}
\alias{transform_features}
\title{Transform feature abunadnce table (modified from Maaslin2)}
\usage{
transform_features(features, transform = "NONE", pseudo_count = 0)
}
\arguments{
\item{features}{feature-by-sample matrix of abundances (proportions or
counts).}

\item{transform}{transformation method.}

\item{pseudo_count}{pseudo count to be added to feature_abd..}
}
\value{
transformed abundance table.
}
\description{
Transform feature abunadnce table (modified from Maaslin2)
}
