% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transformFeatures}
\alias{transformFeatures}
\title{Transform feature abunadnce table}
\usage{
transformFeatures(features, transform = "NONE", pseudo.count = 0)
}
\arguments{
\item{features}{feature*sample abundance table.}

\item{transform}{transformation method.}

\item{pseudo.count}{pseudo count to be added to abundances before normalization.}
}
\value{
transformed feature*sample abundance table.
}
\description{
Transform feature abunadnce table
}
