% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{TSS}
\alias{TSS}
\title{TSS normalization (modified from Maaslin2)}
\usage{
TSS(x)
}
\arguments{
\item{x}{vector of abundance to be normalized.}
}
\value{
normalized vector of abundance.
}
\description{
TSS normalization (modified from Maaslin2)
}
\keyword{internal}
