% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize.continuous.discover}
\alias{visualize.continuous.discover}
\title{Title}
\usage{
visualize.continuous.discover(pc.graph, membership.loading,
  size.communities, plot.size.cutoff)
}
\arguments{
\item{pc.graph}{the full pc network constructed from correlated PCs}

\item{membership.loading}{membership of PC loadings from community discovery}

\item{size.communities}{ordered (largest to smallest) size of the identified communities}

\item{plot.size.cutoff}{cluster size cutoff (for cluster to be included in the visualized
PC network.)}
}
\value{
an invisible list of the subsetted network and memberships (to reproduce the plot)
}
\description{
Title
}
