% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_continuous_discover.R
\name{visualize_continuous_discover}
\alias{visualize_continuous_discover}
\title{Visualization of the clustered network for the continuous.discover function}
\usage{
visualize_continuous_discover(
  graph_pc,
  membership_loading,
  size_communities,
  plot_size_cutoff,
  short_names,
  output
)
}
\arguments{
\item{graph_pc}{the full pc network constructed from correlated PCs}

\item{membership_loading}{membership of PC loadings from community discovery}

\item{size_communities}{ordered (largest to smallest) size of the identified 
communities}

\item{plot_size_cutoff}{cluster size cutoff (for cluster to be included in 
the visualized PC network)}

\item{short_names}{shorter names of the loadings}

\item{output}{output file name}
}
\value{
an invisible list of the subsetted network and memberships (to 
reproduce the plot)
}
\description{
Visualization of the clustered network for the continuous.discover function
}
\keyword{internal}
