% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.batch.R
\name{adjust.batch}
\alias{adjust.batch}
\title{Zero-inflated Empirical Bayes Adjustment of Batch Effect in Feature Abundance Data}
\usage{
adjust.batch(feature.count, batch, formula.adj = NULL, data.adj = NULL,
  zero.inflation = TRUE, pseudo.count = 0.5, diagnostics = TRUE,
  verbose = TRUE)
}
\arguments{
\item{feature.count}{Feature x sample matrix of feature abundance (counts)}

\item{batch}{Variable indicating batch membership}

\item{formula.adj}{Formula indicating additional covariate to adjust for
in the batch correction model}

\item{data.adj}{Data frame for covarites to adjust for.}

\item{zero.inflation}{Flag for whether or not a zero-inflated model should be run.
Default to TRUE (zero-inflated model). If set to FALSE then vanilla ComBat
(with parametric adjustment) will be performed.}

\item{pseudo.count}{Pseudo count to add to the count table before log-transformation. Default
to 0.5.}

\item{diagnostics}{Flag for whether or not diagnostic figures should be generated.
Deafault to TRUE.}

\item{verbose}{Flag for whether or not verbose modelling information should be printed.
Default to yes.}
}
\value{
A feature x sample matrix of adjusted feature abundance.
}
\description{
Zero-inflated Empirical Bayes Adjustment of Batch Effect in Feature Abundance Data
}
