% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_batch.R
\name{adjust_batch}
\alias{adjust_batch}
\title{Zero-inflated empirical Bayes adjustment of batch effect in compositional
feature abundance data}
\usage{
adjust_batch(feature_abd, batch, covariates = NULL, data, control)
}
\arguments{
\item{feature_abd}{feature-by-sample matrix of abundances (proportions or
counts).}

\item{batch}{name of the batch variable. This variable in data should be a
factor variable and will be converted to so with a warning if otherwise.}

\item{covariates}{name(s) of covariates to adjust for in the batch correction
model.}

\item{data}{data frame of metadata, columns must include batch and covariates
(if specified).}

\item{control}{list of control parameters. See details.}
}
\value{
feature-by-sample matrix of adjusted feature abundance.
}
\description{
Zero-inflated empirical Bayes adjustment of batch effect in compositional
feature abundance data
}
