% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_lm_meta.R
\name{check_covariates_random}
\alias{check_covariates_random}
\title{Check random covariates}
\usage{
check_covariates_random(data_covariates, batch)
}
\arguments{
\item{data_covariates}{data frame of random covariates.}

\item{batch}{batch variable.}
}
\value{
vector of indicators per batch for if/which random covariates can be 
fitted within the batches
}
\description{
Check random covariates
}
\keyword{internal}
