% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_maaslin_meta.R
\name{check_exposure}
\alias{check_exposure}
\title{Check exposure variable consistency across batches}
\usage{
check_exposure(exposure, batch)
}
\arguments{
\item{exposure}{Vector of the exposure variable.}

\item{batch}{Factor vector indicating batch/study membership.}
}
\value{
A logical vector indicating if the exposure has variance (>1 level) 
        within each batch.
}
\description{
Check exposure variable consistency across batches
}
\keyword{internal}
