% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_adjust_batch.R
\name{construct_design}
\alias{construct_design}
\title{Construct a non-intercept design model matrix given a metadata data frame}
\usage{
construct_design(data)
}
\arguments{
\item{data}{metadata data frame}
}
\value{
design matrix
}
\description{
Construct a non-intercept design model matrix given a metadata data frame
}
