% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.discover.R
\name{continuous.discover}
\alias{continuous.discover}
\title{Main function for continuous structure discovery}
\usage{
continuous.discover(feature.abd, batch, data, normalization = "TSS",
  transform = "AST", pseudo.count = 0.5, var.perc.cutoff = 0.8,
  cor.cutoff = 0.5, cluster_function = igraph::cluster_optimal,
  plot.clustered.network = TRUE, plot.size.cutoff = 2,
  diagnostics = FALSE, verbose = TRUE)
}
\arguments{
\item{feature.abd}{feature*sample matrix of feature abundance (counts preferred if transform
is specified as LOG).}

\item{batch}{name of the batch variable.}

\item{data}{data frame of metadata, must contain batch.}

\item{normalization}{normalization parameter.}

\item{transform}{transformation parameter.}

\item{pseudo.count}{pseudo count to add to the count table before log-transformation. Default
to 0.5. This should be changed for relative abundance (will automatically set to half of minimal
non-zero value if not specified).}

\item{var.perc.cutoff}{percentage variance explained cutoff to choose the top PCs.}

\item{cor.cutoff}{correlation cutoff to construct edges for the PC network.}

\item{cluster_function}{function used to perform network community structure discovery.}

\item{plot.clustered.network}{should the clustered PC network be visualized? Deafault to TRUE.}

\item{plot.size.cutoff}{cluster size cutoff (for cluster to be included in the visualized
PC network.)}

\item{diagnostics}{should diagnostic plots be generated? Deafault to FALSE.}

\item{verbose}{should verbose modelling information should be printed? Default to TRUE.}
}
\value{
a list of continuous structure discovery results, including the following components:
consensus.loading: matrix of the identified consensus loadings, each column for one identified
PC loading cluster.
scores: matrix of the identified continuous scores for all samples, each column for one identified
PC loading cluster.
membership: list of PC loading members for each identified cluster.
mat.vali: matrix of validation correlations of the identified consensus loadings, each column for
one identified PC loading cluster.
clustered.network: list for the the constructed PC network and community discovery results
mat.cosine all-versus-all matrix of cosine coefficients between PC loadings used to construct the network
}
\description{
Main function for continuous structure discovery
}
