% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.meta.R
\name{lm.meta}
\alias{lm.meta}
\title{Main function for batch effect adjustment}
\usage{
lm.meta(feature.count, batch, exposure, covariates = NULL,
  covariates.random = NULL, data, normalization = "TSS",
  transform = "AST", analysis_method = "LM", rma_method = "REML",
  forest.plots = TRUE, directory = "./MMUPHin_lm.meta/",
  verbose = TRUE)
}
\arguments{
\item{feature.count}{Feature x sample matrix of feature abundance}

\item{batch}{Name of the batch variable}

\item{exposure}{Name of the primary exposure of interest variable}

\item{covariates}{Additional covariates for adjustment in individual regressions}

\item{covariates.random}{Group indicator for random effects}

\item{data}{Data frame for metadata.}

\item{normalization}{Normalization parameter for Maaslin2}

\item{transform}{Transformation parameter for Maaslin2}

\item{analysis_method}{Model parameter for Maaslin2}

\item{forest.plots}{Flag for whether or not forest plots figures should be generated.
Deafault to TRUE.}

\item{directory}{Directory for Maaslin output/generated forest plots}

\item{verbose}{Flag for whether or not verbose modelling information should be printed.
Default to yes.}
}
\value{
a list
}
\description{
Main function for batch effect adjustment
}
