% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{LOG}
\alias{LOG}
\title{LOG transformation (modified from Maaslin2 and is different)}
\usage{
LOG(x)
}
\arguments{
\item{x}{vector of abundance to be transformed.}
}
\value{
transformed vector of abundance.
}
\description{
LOG transformation (modified from Maaslin2 and is different)
}
\keyword{internal}
