% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_lm_meta.R
\name{rma_wrapper}
\alias{rma_wrapper}
\title{Wrapper for fitting fixed/random effects meta-analysis model using metafor}
\usage{
rma_wrapper(maaslin_fits, method = "REML", output = tempdir(),
  forest_plot = NULL, rma_conv = 1e-06, rma_maxit = 1000,
  verbose = TRUE)
}
\arguments{
\item{maaslin_fits}{list of Maaslin2 result data frames, outputted from 
Maaslin2_wrapper.}

\item{method}{meta-analysis model to run, options provided in metafor::rma.}

\item{output}{directory for the output forest plots.}

\item{forest_plot}{logical. should forest plots be generated for 
the significant associations.}

\item{rma_conv}{rma threshold control.}

\item{rma_maxit}{rma maximum iteration control.}

\item{verbose}{should verbose information be printed.}
}
\value{
a data frame recording per-feature meta-analysis association results.
(coefficients, p-values, etc.)
}
\description{
Wrapper for fitting fixed/random effects meta-analysis model using metafor
}
\keyword{internal}
