% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/helpers_maaslin_meta.R
\name{shorten_name}
\alias{shorten_name}
\title{Utility for shorter names
Useful when plotting per-feature figures where feature names could be cutoff}
\usage{
shorten_name(x, cutoff = 15)

shorten_name(x, cutoff = 15)
}
\arguments{
\item{x}{vector of names}

\item{cutoff}{number of maximum string length before start cutting off the
middle}
}
\value{
vector of new names with .. replacing the middle part if name is
longer than cutoff
}
\description{
Utility for shorter names
Useful when plotting per-feature figures where feature names could be cutoff

Utility to shorten names for plotting
}
\keyword{internal}
